pyproject.toml
[tool.poetry]
name = "rizzume"
version = "0.1.0"
description = ""
authors = ["author <author@email.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
groq = "^0.8.0"
python-docx = "^1.1.2"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.6"
pyright = "^1.1.365"
pylint = "^3.2.2"
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
